Any user who can submit a job can cause the Schedd to fail to read the job_queue.log
This fix is to translate all empty values to UNDEFINED immediately as
they appear. It is arguably wrong to put information about UNDEFINED
at the log level, but the semantics make sense and this the smallest
fix that closes the issue.

Issue:

$ condor_qedit 1.0 AB ""
Set attribute "AB".

$ grep AB spool/job_queue.log
103 1.0 AB

And on a schedd restart or log compression:

ERROR "Error: bad record with op=103 in corrupt logfile" at line 1048 in file
classad_log.C