Remove itoa.
It was actually buggy in that it was using a static buffer,
which is a no-no in a multi-threaded program.  In fact, all
uses of it could instead use the original const char *
(instead of converting it to an int and then back again), so
just pass the char * around instead.

Signed-off-by: Chris Lalancette <clalance@redhat.com>