Allow custom deltacloud fields in ULOG_ events.
There are certain pieces of information that only condor
knows about a deltacloud instance.  These include, but are
not limited to the public and private IP addresses of the
instance, the available actions to be taken on the instance,
and the provider ID for the instance (which can be different
from the name we supplied).  However, this information
is needed by the deltacloud frontend UI to display to the
user.

This patch stores these attributes in the deltacloud job
classad.  When the gridmanager is properly configured, it
will output this information to the event log, and from
there it can be propagated to the UI.

Signed-off-by: Chris Lalancette <clalance@redhat.com>