Handle deltacloud jobs that start in STOP state.
Some clouds create instances in a STOPPED state, and
it is then up to the user to start them.  This is not
currently handled 100% properly in dcloudjob.cpp.
For a cloud that does this, the sequence looks like:

GM_CREATE_VM -> GM_START_VM -> GM_SAVE_INSTANCE_ID
 -> GM_SUBMITTED

From here, the next correct state would be GM_PROBE_JOB.
However, because we weren't tracking the fact that this
job was created STOPPED, we were then doing:

GM_DONE_SAVE -> GM_STOPPED

which would kill off the instance.  To fix this, we track
the fact that this instance was created STOPPED, and only
go into GM_DONE_SAVE if we failed somehow.

Signed-off-by: Chris Lalancette <clalance@redhat.com>